Refactored image uploads, and automatically fill caption using the image metadata.
There were use cases from createMediaFromFile function that duplicated logic in mediaUpload. All direct usages of createMediaFromFile were removed, leading to the removal of duplicate logic .

mediaUpload was improved to read the caption field from the uploaded images, so the captions are automatically filled. The server fills the captions using the image metadata.